Move the last of the files into sub-packages
And delete the make.py file - move the header about datargs into
the __init__.py
