+++ shogun/__init__.py
@@ -1,3 +1,8 @@
-from .make import parse, arg, make_parser, argsclass
+# Initial implementation courtesy of https://github.com/roee30/datargs
+# under the MIT license - maintained in this repository at LICENSE-datargs
 
-__all__ = ["parse", "arg", "make_parser", "argsclass"]
+from .argparse_.args import arg
+from .argparse_.argsclass import argsclass
+from .argparse_.parse import parse
+
+__all__ = ["arg", "argsclass", "parse"]
